Refactor IntervalCollections to remove unused code, and simplify types (#20046)
Interval collections used to share code with share map. This has not
been the case for a long time, however many of the shared pieces were
just copied over when they were split. This leads to overly generalized
types and implementations in interval collections which add complexity
to the code that isn't necessary. This change removes a bunch of unused
code and removes generalization from many of the types. There should be
no functional changes to anything actually used. There is more
simplification possible here but doing it piecemeal will make it easier
to review and reduce the chance for mistakes.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>