Improve type inference for inserted SharedTree content when the schema is a union of objects (#20091)
## Description

This PR adds some logic to examine the required fields in the schema
when inserting an object. It correctly disambiguates inserted objects
more frequently. For example, if a schema allows for either of two
object types to be present, we now look at the fields of the inserted
objects and can tell which type is being inserted if the fields that are
present satisfy one type but not the other.

This is only an incremental improvement in this space - there are many
more scenarios that could be improved with stronger disambiguation.
Those will be tackled in the future (see the new TODOs in the file).