make deleteChildSummarizerNode non-optional, add better API in ChannelCollection to delete children (#20079)
Expose a way to delete a node in channel collection.
It will be used in following cases:

When GC tells ChannelCollection when some of its routes are not used. (current usage)
in future scenarios where custom solutions based on ChannelCollection would manage lifetime of their nodes (outside of GC sub-system).