Adding baseVersionForMinCompat for min compat version testing (#20060)
Every time we release, the base version in which N-i calculations are
based changes. This causes issues at testing specially at bumping the
release version. By hard coding the version in which these calculations
are made, we decoupled the min version testing from the release process,
although we will need to update this hard coded value shortly after the
release as after work to make sure that min versions are going to work
with the new version. I really believe this is the best attempt so far
to test this functionality, given that it's a middle ground between both
contending arguments of making the tests deterministic and at the same
time not requiring too much of maintenance.