feat(fluid-build): Add incremental build support for biome tasks (#20173)
Adds basic incremental build support for biome tasks to fluid-build.

The implementation does not parse biome config files to find files to
ignore. Instead, it uses `git ls-files` to enumerate all files in the
package folder, along with any biome config files in the tree, and uses
those as the input/output files. This means it includes more files than
it should in its cache, and will be invalidated by edits to any files in
a project. While this seems like overkill right now, when we use biome
for all formatting as a prettier replacement, this approach should "just
work" without further code changes and we can add ignore support later
if needed.