build(routerlicious-driver): use fluid-tsc (#19814)
Semi-automated conversion to fluid-tsc.

1. Axios workaround via re-export from .cts
2. One of the test in mapWithExpiration.spec.ts is modified to run in
CJS module only. This test verifies the behavior of arrow functions
regarding their 'this' context. In ECM, the 'this' value inside an arrow
function is 'undefined', whereas in CJS, it is '{__esModule: true}'.
Therefore, the test exempts the case where 'this' is 'undefined' from
the assertion.

---------

Co-authored-by: Sonali Deshpande <48232592+sonalideshpandemsft@users.noreply.github.com>