server: provided user was not an azure user fixes (#20150)
## Description

This PR _begins_ the process for fixing the `Provided user was not an
"AzureUser"` bug, which was caused by #19000. However, we cannot just
revert #19000 because it was implemented to resolve privacy and data
residency concerns.


![image](https://github.com/microsoft/FluidFramework/assets/14362200/948f8823-895e-4945-8e4a-b84dd6cdbfc7)

[ADO
7422](https://dev.azure.com/fluidframework/internal/_workitems/edit/7422)

Included here is part 1 & 2 of the code changes needed to resolve the
incident. Configs added are pretty self-explanatory:

- `scrubUserDataInGlobalCheckpoints`: this is going to stay enabled
- `scrubUserDataInLocalCheckpoints`: this is going to eventually be
disabled once work lands to enable a 24hr TTL in local db
- `scrubUserDataInSummaries`: this is going to stay disabled, but
including for completeness

Otherwise, the main change is to validate the quorum member information
on Scribe boot in an effort to recover from a global DB quorum read by
loading checkpoint from the summary instead.

### Follow-up Work

Follow-up work will include making some changes to how we create/update
the Checkpoint DB TTL index. Once that is complete we can stop scrubbing
local checkpoint user data.