Fix internalSchema() (#20223)
The root cause of the issue was that getRequestedVersion() was returning
2.0.0-rc.2.0.0 to this: (!!!)
baseVersion === '2.0.0-rc.1.0.5' && requested === -1
The reason why we returned that is because internalSchema() returned
`^2.0.0-internal.8.0.0` to such a case which is deeply wrong. Such a
request to resolveVersion() will produce an array with all versions
ahead of that one and return the latest. Probably the code was looking
for the latest 2.0.0-internal.8.0.0 patch.
This hints me that there is a deeper issue here in how we manage these
cases. Currently hard-coding the desired versions for testing and
probably unblock the patch release for RC1.