Make attribution policy factory exposure consistent (#20226)
## Description

This exposes property-based attribution policy factories from
merge-tree. These two functions were intended to have the same
visibility as `createInsertOnlyAttributionPolicy`, but were never
exported. Since original definition, we've changed the intent of the
\@alpha tag from API extractor. It makes more sense to leave both of
these functions internal (as `createInsertOnlyAttributionPolicy` is), so
I've updated the attribution section of the README as well to be more
accurate.

Co-authored-by: Abram Sanderson <absander@microsoft.com>