build: Add 'api' to 'compile' task definition (#20247)
Now that we are doing API trimming in packages, the `api` task needs to
be considered part of the core package compilation. The task definition
is updated to reflect this.

Practically speaking this will mean that each package will invoke its
`api` task, whereas before, the task was invoked by another package that
depended on it, because the `compile` task _did_ include `^api` in its
task chain already.

I also updated the CI pipelines for protocol-defs and common-defs, since
they now use fluid-build and should be triggered when changes are made
to the config file.