Improve ST proxy hydration (#20246)
## Description

This overhauls much of the mechanism for hydrating ST proxies. There is
a markdown file included in this PR which discusses the overall scheme.
The high level changes of note are:

* Proxy hydration is now more robust (in particular, a bug was fixed in
which hydration failed when inserting multiple consecutive non-primitive
nodes in a list at once).
* Proxy hydration is now lazier. Proxies are associated with their
respective anchor nodes immediately upon insertion, but the flex nodes
are no longer generated eagerly. Instead, they are generated when their
corresponding proxy is next read.
* The bookkeeping (e.g. maps that associate nodes across different API
layers) is now mostly done in a single file (proxyBinding.ts) and no
longer relies on the creepy `onNextChange` event, which has been
removed. Instead, the mappings are routed through anchor nodes which are
created when the node is inserted.
* Another suite of tests has been added which checks that proxies can be
read (after insertion) without having first read their parents.

co-authored by @taylorsw04