improvement(client): IDocumentDeltaConnection.submitSignal typing BREAKING CHANGE (#20251)
`IDocumentDeltaConnection.submitSignal` is not known to actually be used
externally. Internally only string is used and we'll use the most strict
type until there is need to do something else.

`Jsonable<>` content is the least restrictive type that could be used,
but that type filter currently has defects; so, keep with the simple
strict string.

Driver `emitMessages` is now broken down into the two uses with specific
and accurate typing.