feat(build-cli): Add `modify fluid-imports` command (#20006)
Adds a `modify fluid-imports` command that can be used to rewrite the
Fluid imports of a project to use the correct subpath import based on
the API level of a given import. The mapping data only includes a few
packages so that's all that's fixed up.

I ran the command on the aqueduct project as part of testing. This is
the command I used:

```shell
flub modify fluid-imports -v --onlyInternal
```

If I omit `onlyInternal` then it properly splits the single import
statement into some alpha and some internal, which triggers our lint
rule preventing alpha imports. Good news!

### Known issues

- Default imports are not handled. If encountered a warning will be
shown and the import declaration will be skipped.