fix: Do not attempt to hydrate ST proxies in initial tree (#20326)
## Description

This fixes a bug in which proxies in the initial tree given to a
SharedTree are hydrated even when the initial tree function is unused
(e.g. when loading an existing document that already has an initial tree
with the current schema). This hydration shouldn't occur because in that
case the proxies produced by the initial tree function are never used
and are never inserted into the tree. For now, proxy hydration is simply
disabled for proxies in the initial tree.