build: Pre-apply upcoming policy changes (#20254)
Pre-applies some upcoming policy changes to the repo, most notably the
requirement for an empty line between the copyright header and any other
text/code. It also adds a temporary exclusion for the generated type
tests since they are generated during build and thus their contents
depends on build-tools, and is guaranteed to have the header. Regardless
we can remove this once all release groups and packages use 0.35.0+ of
build tools.