Refactor compat test code to be able to better control versions for corss-compat tests (#20323)
This is mechanical refactoring - there should be no changes to observed behavior.
If you see where it's not the case, please flag it.

Main purpose of this refactoring is to create an opportunity for genCrossVersionCompatConfig() to control what versions are used in cross-version testing. Currently there are two problems as I see them:
1. The values in defaultCompatVersions have different meaning depending on type of the test. genCrossVersionCompatConfig and the rest of the pipeline dealing with cross-version type of tests (i.e. "TestObjectProviderWithVersionedLoad" type) treats -1 as previous major, vs. other tests treat it as previous minor.
2. genCrossVersionCompatConfig does not control nor have mechanism to precisely tell the version to be used. I.e. it has no way to say - use previous minor. That's because getVersionedApis() will treat -1 as major for cross-compat tests (passing adjustMajorPublic = true to all APIs), but will treat -1 as minor (not passing 3rd argument to various APIs, which means adjustMajorPublic = false)

This change moves decision making up the stack (at least for cross-compat tests) from getVersionedApis() into genCrossVersionCompatConfig() which now conveys precise version to use, and thus has all control to prescribe any version to use.

Overall, I think all this code should move to a model of passing exact versions. I.e. only configList (lazy promise) should manipulate with -1 kind of math, but the output of it should be list of precise versions to use, and the rest of the code neither should do any math nor assume that -1 has some specific meaning. I've stopped short of achieving this goal, as there is more risk I'll break something, and my need is to change cross-version tests (in future PRs).