Add gcVersion to sendGCUnexpectedUsageEvent (#20023)
It's important to know which GC Version is used by the code logging
critical GC errors. The value logged will be the version _in effect_,
not the version in the base snapshot. This is sufficient.

For instance, if the code knows about GC Version 3 and hits a Tombstone
error, then we know that the GC Data in the document must also be for GC
Version 3 (otherwise it would have dropped all tombstones on load, and
rerun full GC).

But if the code only knows about GC Version 2 and hits a Tombstone
error, we should ignore it, since there were known issues with GC
Version 2.