build(api-markdown-documenter): Use eslint config from npm (#20369)
Because the api-markdown-documenter project depends directly on the
eslint config, unlike the rest of the repo, it makes it challenging to
update the lint config _without_ making changes to
api-markdown-documenter. See #20363 for an example. Eventually I would
like to get the whole repo to use file: deps for the lint config, but
when it's only one project, it makes things more difficult.

I have also noticed that auto-fixers, especially those in the "import"
eslint plugin, do not work as expected in this project, which I suspect
is related to the file: dependency. I've observed the same behavior when
trying file: dependencies in other release groups.