fix(build-common): Update base API-Extractor configs to include `tsdocMetadata` to ensure cross-package trimming is handled correctly (#20381)
API-Extractor uses the presence of this file in an npm package to
determine whether or not its API members should be considered when
making decisions about reports and rollups in consuming packages.
This includes determining whether or not release tags on external
members should be respected when making trimming decisions.
We need to ensure this is enabled so that re-exports of API members
across package boundaries are correctly trimmed based on their release
tags.

Note: secondary config files are unused. They probably should have tsdocMetadata enabled, but will be deleted soon anyway.