Implement explicit runtime version rejection mechanism. (#20375)
This is mostly a back-up mechanism to kick out old runtimes that are not compatible with the latest runtime.
It's backup in a sense that we should prefer an explicit mechanism by adding new property to a list of features supported / required by runtime.

However, we could find bugs much later in time, where multiple runtime versions shipped claiming they understand new capabilities. In simple cases we can simply rename / add new property to signify that only latest runtime version actually understands it. But in more nuanced cases (where some old runtime versions are broken, and some old runtime versions are not), this mechanism will be insufficient.

I prefer here to check exact version, and if needed, list 10 versions, as opposed to do math on them (i.e. do something like >=). The latter is a bit risky as we keep changing our version schema encoding, and I'm not confident we can predict how these kinds of math comparisons will work out. Plush do not want to take a dependency on semver package.

The amount of new code here is relatively small, and it's dormant until we start using it in some future. Most changes are UTs.

This work builds in previous work:
- https://github.com/microsoft/FluidFramework/pull/20174
- https://github.com/microsoft/FluidFramework/pull/20297