Add testing prior minor to our cross-version tests (#20353)
Ensure that we at least test (prior minor, current version) mix (both directions) when running cross-version tests, i.e. ensure that document format (ops, summaries) is compatible across two closest minor releases.

This change by itself does not add that much coverage but creates possible framework for future changes.
Current tests only validate that with a mix of 1.x and latest 2.x. This is good, but we might miss cases that way. Also, tests might be disabling cross-compat tests altogether because they are testing recently added features - we would love to move toward a direction where such tests only disable parts of cross-version tests - the part that has 1.x in it but enable the other part - testing N / N -1. We can even introduce a new mode in addition to "Full Compat" that enables N / N-1 in next release, such that developers can add tests and not need to go back to enable partial compat mode (all TBD, and not prescribed here).

It will increase time to run tests, as it doubles number of cross-version pairs. That said, it adds only 2 configs to already 13 configs we are testing in Full compat mode, and not all tests use Full compat mode.
I've tested it on my box, running only end-to-end tests:
1. Without this change: 2m 55 seconds
2. With this change: 3m 18 seconds

So that's 13% increase (based on one sample)
Relative %% will be lower if considering running all the tests (as there are more tests that are not affected by this change).