Use correct clientSequenceNumber in mocks (#20368)
### Description

Currently, the attribution fuzz test validation relies on the
`clientSequenceNumber` starting at 0. This does not happen in
production.

Also, since the "content" it matches on is string comparison, any
different ordering of properties in ops will cause the validation to
fail even though parsing to JSON would produce equal objects.


[AB#7512](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7512)

### Reviewer Guidance

The only changes to the snapshot `.json` files are increasing each
`clientSequenceNumber` by 1.