fix(datastore): Fix bundle size regression caused by lodash import changes (#20421)
This PR fixes the bundle size regression that was caused by the changes
in the lodash import syntax.

The syntax was initially changed to accommodate the new node 16 module
resolution (https://github.com/microsoft/FluidFramework/pull/18826). The
syntax it was changed to caused the entire lodash library to be loaded.
This PR fixes this by replacing cloneDeep with structuredClone