compressor: Creation ranges include which IDs were generated as local (#20424)
This PR adds information to ID creation ranges to describe which IDs in
the range were generated as local IDs on the originating client. This
information is currently unused but will be used in the future to
support building incremental stashed state. Specifically, the local
ranges will be consumed during application of stashed state to re-create
the normalization maps inside the compressor rather than having to store
the serialized tip state of the compressor in the stashed state.

This changes the persisted format of the compressor ranges (and thus
runtime ops containing them). This will break support for any documents
made by customers who are using the new SharedTree (as it is the only
feature using the compressor) and have stashed ops, which is currently
not possible through the public API.