Await promises from runWithRetry (#20436)
## Description

This PR fixes a regression that was introduced in
https://github.com/microsoft/FluidFramework/pull/20243.
https://github.com/microsoft/FluidFramework/pull/20243 added the use of
`runWithRetry` to historian which caused a bug in the ephemeral
containers workflow. The bug was that the `isEphemeral` flag was not
always set to `true` as the async call to set/get it from the cache was
not awaited. This PR adds an `await` to fix the issue.