tree: Add missing schema agnostic indexed field traversal to flex tree, use in proxy binding (#20440)
## Description

Flex-tree is supposed to provide good APIs for working with the tree in
non-schema aware code.
While it had iteration APIs, it was missing indexed access. This fixes
that omission, and uses it (instead of down casting to the schema aware
api) in proxyBinding's demandSpine.

This also documents some TODOs to cleanup demandSpine.