feat(tree): Add the ability to specify stable identifiers to object fields (#20372)
Adds the ability for users to specify explicit field identifiers for use
in the persisted tree format, which may disagree with the developer API
model.

The API changes here reflect the design outlined
[here](https://microsoft.sharepoint.com/:w:/t/FluidFramework/EUxBXfoEpehLldqMGUaDDXYB6BZI54h2nn6_TtH742quGg?e=BANqa6)
(Microsoft internal).

Noteworthy changes to the `simple-tree` layer:
* Users may specify `StoredKey`s in object node schemas. These keys
affect the persisted format of the tree, but do not affect the
developer-facing property APIs. This is particularly useful for
supporting schema refactoring scenarios while maintaining compatibility
with existing data.
* The `simple-tree` layer now has to differentiate / map between the two
kinds of keys, but the `flex-tree` layer _strictly_ operates in terms of
stored keys. As a result, the "view" keys used in the `simple-tree`
developer model become effectively session-local and ephemeral.

I have attempted to document the distinctions here clearly in the
sourcecode docs. Any feedback on the efficacy of those docs would be
greatly appreciated.

---------

Co-authored-by: Craig Macomber (Microsoft) <42876482+CraigMacomber@users.noreply.github.com>