feat(build-tools): `fluid-imports` to read API levels from packages (#20437)
when not supplied by data file. Default data file has been removed in
favor of the real dependencies.

complimentary features:
- add multiple project processing to reduce overhead and allow scripts
to run against a bunch of packages with disparate config files (just
needs one to match). Script must still be run local to package for
ts-morph to recognize imports. This is good as it helps to avoid problem
where multiple projects are given and should be working off of different
versions of Fluid packages.
- add package filter regex
- revise logging