tree: Centralize ArrayNode logic in one file (#20462)
## Description

Currently code for node kinds is organized by what part of the
implementation logic it is, not by what kind of node it is.

This change reverses that for ArrayNode (with the expectation that if we
do this for ArrayNode, the others will follow)

This has a few benefits:

- Saves on dynamic dispatching on NodeKind in some places (ex:
copyContent).
- Allows implementation details (like rawFieldNode) to be non-exported.
- Makes it practical to refactor how the implementation of a given node
kind works without impacting others (for example if we wanted to make
one node kind proxyless).
- Makes it more clear what is needed to add a node kind and easier to
fork a node kind (when needed for compat reasons).

This does not fully centralize the array node kind logic since a few
things are still scattered. These include at least:

- The node kind enum member. It may be practical to remove this in the
future if another down casting/dispatch option is added (like a common
base class).
- The schema factory methods to create array nodes (all the real logic
has moved, but the entry points still live on schema factory: this seems
fine).
- toMapTree's compatibility and hydration logic (there are plans to
address this).