Make some public interface members readonly (#20422)
## Description

Remove mutability from a few places in the public API it does not appear
to be needed.

The change to IValueChanged seems like a bug fix since the same object
is provided to all registrants for the event, so they modify the object
it would impact other event receivers which seems like bad coupling. I
Suspect the other changes are similar in nature: I don't think mutation
of any of these was intended to be supported or meaningful, but I could
be wrong about this.

## Breaking Changes

This is a breaking change, since some users of these interfaces could be
modifying them, and this could break.