Change the order of loggers in UT infrastructure, to get key properties logged to CI whenever error is logged into telemetry (#20408)
As I was working on PR #20353, I had really hard time figuring out why it fails in the lab. All tests were passing on local box.
After a bunch of struggling, I've opened https://dev.azure.com/fluidframework/internal/_queries/edit/7612/?triage=true that describes that problem of our test system that does not log key properties that would make my investigation take minutes, instead of hours.

As result, this PR, which is a partial solution, but that was sufficient for me to figure out problem in my code.

The core of the issue - we install tracker above key loggers that log valuable properties / events. Loggers are organized like a chain. All loggers that follow tracker in that chain report to void - we do not see their properties in logs captured (for example) for unexpected (by test) error telemetry events.
Some of these loggers carry very important information about test or environment that makes it much easier for developers to reason about errors or environment.