Signal tests on server (#20516)
## Description

We noticed there currently aren't any unit tests for signals on the
server. This was probably overlooked due to the simplicity of signal
logic and usage. However, with the future added complexity of targeted
signals (#19519) we thought it would be smart to invest in backfilling
the missing signal test coverage before moving forward.

ADO Task:
https://dev.azure.com/fluidframework/internal/_workitems/edit/7657
Next Steps: #19519  

cc: @sashasimic 

## Reviewer Guidance
Need verification on: 
- Correct testing pattern/style
- That this is the correct file to place these tests