tree: Remove child change generic parameter from sequence changeset (#20518)
## Description

Sequence field had a generic parameter for child changes. This was
previously used to allow for `NodeChangeset` children in production and
other types such as `TestChange` in testing. This PR removes this
generic parameter and requires that child changes of a sequence are
always represented as a `NodeId`. In testing a `TestChange` can either
be embedded in the `NodeId` or it can be associated with it out of line
using `ChangesetWrapper`.