Repro offline initial sequence number mismatch (#20396)
[AB#7667](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7667)

In the work to do
[AB#7588](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7588),
I found that the core assumptions I had about
`DeltaManager.initialSequenceNumber` in the runtime layer was wrong. My
understanding was that `DeltaManager.initialSequenceNumber` pointed to
the sequence number of the base snapshot. This was broken in an offline
scenario that wanted to prevent the `DeltaManager` from processing extra
ops. This PR fixes that so that the `DeltaManager` still skips
processing extra ops and has its initial sequence number point to the
sequence number of the snapshot the container loaded from.