Add gc test for data virtualization (#20185)
[AB#7412](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7412)

Add gc tests for data virtualization.

The scenario here is the one I was most concerned about, basically does
GC data passed down from the summary still work. Since we designed GC to
work in "unrealized" datastores. The answer is yes, as this test proves
such.

While I was testing this, I discovered that the call to
`FluidDataStoreContext.isRoot` occurs during summarization for all
datastores including virtualized datastores. This means that our
"virtualization" would be moot.

I introduced a new change that would make it so that our virtualized
datastores would not be loaded during summarization.