Export ISnapshotInfo and create conversions (#20530)
Part of a bigger effort for Data virtualization and offline, this is the
reference PR: https://github.com/microsoft/FluidFramework/pull/20342


[AB#7670](https://dev.azure.com/fluidframework/235294da-091d-4c29-84fc-cdfc3d90890b/_workitems/edit/7670)

Enable conversion from `ISnapshotInfo` to `ISnapshot` as they are
relatively similar. The conversion of `ISnapshot` to `ISnapshotInfo` and
vice versa is important as offline uses `ISnapshotInfo` as it is
serializable and `ArrayBufferLike` doesn't survive serialization.
`ISnapshot` is what data virtualization to load the snapshot.