Allow ID compressor to be enabled in existing files. (#20531)
This work builds on previous work:
- https://github.com/microsoft/FluidFramework/pull/19859
- https://github.com/microsoft/FluidFramework/pull/20174

While prior work gives us opportunity to generate short IDs (this is used, for example, to generate short IDs for data stores and DDSs), the current state of ID compressor settings does not allow us to change them through lifetime of the container.
I.e. if container is created with ID compressor off, then it stays off for duration of container lifetime.

This is not great, and ideally, we want (eventually) all files to use ID compressor, at least for generation of short IDs.
This work makes it possible, allowing off -> delayed mode transition.

It stops short from enabling off | delayed -> on transition, as currently this is not possible with async nature of ID compressor loading and synchronous op processing pipeline. It's possible we can make such transition work with some delay, but more work is required to make it happen.