Fix document schema summarization bug (#20567)
When we create a new document, we incorrectly produce a summary with empty schema, the same way as if we loaded from a document without any schema defined (=== default schema).
This is incorrect, as what we summarize should be equal to current session schema in such case.
This bug happens only if explicit schema management is ON - this explains why this bug was not caught by end-to-end tests.

Also changed out end-to-end tests to test all combinations regarding explicit vs. implicit schema management.