revert preventing of retries for throttling errors (#20543)
This PR reverts the changes made here:
https://github.com/microsoft/FluidFramework/pull/19722

We decided that it would be better to have a more generalized solution
for the throttling errors, where hosts can disconnect the container when
they receive throttling error warnings.

Since there has been no new releases from FF with this change, the
revert seems safe. But would like reviewers to advise as well.