Add "node exists" transaction constraint to SharedTree transaction API (#20606)
## Description

This adds the ability to constrain transactions with "preconditions"
that are evaluated when the transaction is about to be applied. It
allows for the first (and currently only) kind of constraint that
SharedTree permits: a "node existence" constraint which requires that a
given node must be present for the transaction to succeed.