MergeTree: Remove unexposed and redundant  MergeBlock interfaces (#20614)
I'm looking into some changes related to how we layout content in the
merge tree, which will necessitate making changes to the types that make
up the tree. The MergeBlock types which make up the tree have interfaces
and classes internally which is unnecessary and makes refactoring
harder. This change removes the unexposed and redundant MergeBlock
interfaces.

---------

Co-authored-by: Tony Murphy <anthonm@microsoft.com>