Adding verbose logs for a few error events (#20558)
## Description

Adding some verbose logs to track the error handling in case of document
corruption - `markAsCorrupt` function calls `this.context.error` and the
exception goes unhandled somewhere in that flow, causing the service to
restart. So adding these logs to track that in prod as it is not
reproducible locally. Verbose logs will be logged only in pilot
clusters.


![image](https://github.com/microsoft/FluidFramework/assets/22322514/e36223a5-4d24-4ed4-a3b8-b06a74cb3a1c)

---------

Co-authored-by: Shubhangi Agarwal <shuagarwal@microsoft.com>